Move lib files to folder (for namespacing).
